requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
cffi==1.11.5
coverage==4.5.1
docutils==0.14
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
hupper==1.3
Jinja2==2.10
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
opcua==0.98.5
PasteDeploy==1.5.2
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.6
pluggy==0.7.1
py==1.6.0
pycparser==2.19
Pygments==2.2.0
pyramid==1.9.2
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-retry==0.5
pyramid-tm==2.2
pytest==3.8.0
pytest-cov==2.6.0
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
repoze.lru==0.7
six==1.11.0
SQLAlchemy==1.2.12
transaction==2.2.1
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.2
WebTest==2.0.30
zope.deprecation==4.3.0
zope.interface==4.5.0
zope.sqlalchemy==1.0
